@@ -31,18 +26,32 @@
     "react-dom": "^15.0.1",
     "request": "^2.71.0",
     "rimraf": "^2.5.2",
+    "rxjs": "^5.3.1",
     "walk": "^2.3.9",
     "xml-parser": "^1.2.1",
     "yazl": "^2.4.2"
   },
-  "readme": "\n# Bart\n\nDemandware file uploader & script debugger\n\n\n### Usage\n\nAfter installing Bart package, in Atom appear additional item in main menu with same name `Bart`. This item contain three subitems:\n\n* Upload all cartridges - actually upload all cartridges to the sandbox (zip, upload and unzip on SB). Currently only first folder in project will be uploaded.\n* Open settings... - Opens configuration of current sandbox like a hostname, username, password, codeversion and list of files that should be watched. Last option is used to watch files that should be uploaded to SB when change by some other program i.e. on build and minify by gulp/grant, and  is not opened in IDE. Each file should have absolute path and can end by mask i.e. `/.../cartridges/app_core/cartridge/static/default/js/*`\n* Open/close debugger - simply toggle debugger's panel.\n\nTo start debug you heed to set breakpoint first. To do so, just open file set text cursor in required place and press F9. After breakpoint is set open in browser required page. You will see that page is loading too long. So it's time to press `Manual Update` button in debugger panel (last button in row). After that you should see stack trace and list of variables.\nOther buttons should be obvious.\n\nAlso in context menu of file/editor you will see two additional items: upload file and toggle breakpoint.\n\n\nIt's recommended to install also [language-demandware](https://atom.io/packages/language-demandware) for syntax highlight of Demandware files.\n",
-  "readmeFilename": "README.md",
-  "bugs": {
-    "url": "https://github.com/SqrTT/bart/issues"
+  "description": "demandware file uploader and debugger",
+  "devDependencies": {
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^7.0.0"
+  },
+  "engines": {
+    "atom": ">=1.7.0 <2.0.0"
   },
   "homepage": "https://github.com/SqrTT/bart#readme",
-  "devDependencies": {
-    "eslint": "^2.9.0",
-    "eslint-plugin-react": "^5.0.1"
-  }
+  "keywords": [
+    "demandware"
+  ],
+  "license": "MIT",
+  "main": "./lib/bart",
+  "name": "Bart",
+  "optionalDependencies": {},
+  "readme": "\n# Bart\n\nDemandware file uploader & script debugger\n\n\n### Usage\n\nAfter installing Bart package, in Atom appear additional item in main menu with same name `Bart`. This item contain three subitems:\n\n* Upload all cartridges - actually upload all cartridges to the sandbox (zip, upload and unzip on SB). Currently only first folder in project will be uploaded.\n* Open settings... - Opens configuration of current sandbox like a hostname, username, password, codeversion and list of files that should be watched. Last option is used to watch files that should be uploaded to SB when change by some other program i.e. on build and minify by gulp/grant, and  is not opened in IDE. Each file should have absolute path and can end by mask i.e. `/.../cartridges/app_core/cartridge/static/default/js/*`\n* Open/close debugger - simply toggle debugger's panel.\n\nTo start debug you heed to set breakpoint first. To do so, just open file set text cursor in required place and press F9. After breakpoint is set open in browser required page. You will see that page is loading too long. So it's time to press `Manual Update` button in debugger panel (last button in row). After that you should see stack trace and list of variables.\nOther buttons should be obvious.\n\nAlso in context menu of file/editor you will see two additional items: upload file and toggle breakpoint.\n\n\nIt's recommended to install also [language-demandware](https://atom.io/packages/language-demandware) for syntax highlight of Demandware files.\n",
+  "readmeFilename": "README.md",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/SqrTT/bart.git"
+  },
+  "version": "1.7.0"
 }
