@@ -6,8 +6,8 @@ import EventEmitter from 'events';
 
 const noop = function() {};
 
-const StackView = React.createClass({
-    render: function () {
+class StackView extends React.Component {
+    render() {
         if (this.props.thread === 'spin') {
             return <span className='loading loading-spinner-tiny inline-block'></span>;
         } else if (this.props.thread) {
@@ -16,8 +16,9 @@ const StackView = React.createClass({
                     <ul className='list-group'>
                         {this.props.thread.threadInfo.call_stack.map((stack, index) =>
                             <li
+                                key={index}
                                 className={(this.props.selected || 0) === index ? "list-item selected" : "list-item"}
-                                onClick={this.props.onSelectFrame ? this.props.onSelectFrame.bind(this, index) :  noop}
+                                onClick={this.props.onSelectFrame ? () => this.props.onSelectFrame(index) : noop}
                                 title={stack.location.script_path}
                             >
                                 <span>{stack.location.script_path.split('/').pop()}</span>
@@ -32,23 +33,23 @@ const StackView = React.createClass({
             return (<div className="block">No active thread</div>);
         }
     }
-});
+}
 
-const VarView = React.createClass({
-    getInitialState: () => {
-        return {opened: false, pending: false, values: null};
-    },
-    toggleOpened: function () {
-        const self = this;
+class VarView extends React.Component {
+    constructor() {
+        super();
+        this.state = {opened: false, pending: false, values: null}
+    }
+    toggleOpened() {
         this.setState({opened: !this.state.opened});
         if (!this.state.values && this.props.thread) {
             this.setState({pending: true});
-            this.props.thread.getMembers(this.props.path, this.props.selected).then(function (data) {
-                self.setState({pending: false, values: data});
+            this.props.thread.getMembers(this.props.path, this.props.selected).then(data => {
+                this.setState({pending: false, values: data});
             });
         }
-    },
-    render: function () {
+    }
+    render() {
         var pending = '';
         if (this.state.pending) {
             pending = <span className='loading loading-spinner-tiny inline-block'></span>;
@@ -57,10 +58,15 @@ const VarView = React.createClass({
         if (this.state.values && this.state.opened) {
             vals =
                 (<ul className='list-tree'>
-                    {this.state.values.map(value =>
-                        <li className="list-nested-item">
-                            <VarView type={value.type} value={value.value} name={value.name} path={this.props.path + '.' + value.name}
-                                thread={this.props.thread} selected={this.props.selected}
+                    {this.state.values.map((value, index) =>
+                        <li className="list-nested-item" key={index}>
+                            <VarView
+                                type={value.type}
+                                value={value.value}
+                                name={value.name}
+                                path={this.props.path + '.' + value.name}
+                                thread={this.props.thread}
+                                selected={this.props.selected}
                             />
                         </li>
                     )}
@@ -69,8 +75,8 @@ const VarView = React.createClass({
         return (
             <div className="bart_pad_tree">
                 <div className='list-item'>
-                    <span className={this.state.opened ? 'icon icon-chevron-down' : 'icon icon-chevron-right'}
-                        onClick={this.toggleOpened}
+                    <span className={this.props.type.includes('dw.') || this.props.type.includes('Object')  ? 'icon icon-chevron-right' : ''}
+                        onClick={this.toggleOpened.bind(this)}
                     ></span>
                     <span title={this.props.type} >{this.props.name}</span>
                     <span> = {this.props.value}</span>
@@ -79,10 +85,10 @@ const VarView = React.createClass({
             </div>
         )
     }
-});
+}
 
-const VariablesView = React.createClass({
-    render: function () {
+class VariablesView extends React.Component {
+    render() {
         if (this.props.values === 'spin') {
             return <div className='b-bart_variables right'>
                 <span className='loading loading-spinner-tiny inline-block'></span>
@@ -91,8 +97,8 @@ const VariablesView = React.createClass({
             return (
             <div className="b-bart_variables right">
                 <ul className='list-tree'>
-                    {this.props.values.map(value =>
-                        <li className="list-nested-item">
+                    {this.props.values.map((value, key) =>
+                        <li className="list-nested-item" key={key} >
                             <VarView type={value.type} value={value.value} name={value.name} path={value.name}
                                 thread={this.props.thread} selected={this.props.selected}
                             />
@@ -104,29 +110,30 @@ const VariablesView = React.createClass({
             return (<div className="b-bart_variables right">No variables</div>);
         }
     }
-});
+}
 
 
-const MainView = React.createClass({
-    getInitialState: () => {
-        return {thread: null, selectedFrame: 0};
-    },
-    emit: (eventName, data) => {
+class MainView extends React.Component {
+    constructor() {
+        super();
+        this.state = {thread: null, selectedFrame: 0};
+    }
+    emit(eventName, data) {
         this.emitter.emit(eventName, data);
-    },
-    componentDidMount: () => {
+    }
+    componentDidMount() {
         this.emitter = new EventEmitter();
-    },
-    on: (event, fn) => {
+    }
+    on(event, fn) {
         this.emitter.on(event, fn);
-    },
-    onSelectFrame: function (frameNo) {
+    }
+    onSelectFrame(frameNo) {
         this.emit('selectFrame', frameNo);
         this.setState({selectedFrame: +frameNo});
-    },
-    render: function () {
+    }
+    render() {
         return (
-            <div className="b-bart-debugger_panel block" contenteditable="true">
+            <div className="b-bart-debugger_panel block" contentEditable="true">
                 <atom-panel class="b-bart_parent_stack left">
                     <div className="inline-block btn-group">
                         <button onClick={this.emit.bind(this, 'close')} className="icon icon-remove-close btn" title="close debugger F12"></button>
@@ -135,9 +142,8 @@ const MainView = React.createClass({
                         <button onClick={this.emit.bind(this, 'stepover')} className="icon icon-jump-right btn" title="step over F10"></button>
                         <button onClick={this.emit.bind(this, 'stepin')} className="icon icon-jump-down btn" title="step in F11"></button>
                         <button onClick={this.emit.bind(this, 'stepout')} className="icon icon-jump-up btn" title="step out SHIFT-F11"></button>
-                        <button onClick={this.emit.bind(this, 'update')} className="icon icon-repo-sync btn" title="manual update"></button>
                     </div>
-                    <StackView thread={this.state.thread} onSelectFrame={this.onSelectFrame} selected={this.state.selectedFrame}/>
+                    <StackView thread={this.state.thread} onSelectFrame={this.onSelectFrame.bind(this)} selected={this.state.selectedFrame}/>
                 </atom-panel>
                 <VariablesView values={this.state.values} thread={this.state.thread} selected={this.state.selectedFrame} />
                 <div className="clearboth"/>
@@ -145,7 +151,7 @@ const MainView = React.createClass({
             );
 
     }
-});
+}
 
 
 export default class DebuggerView {
