@@ -5,6 +5,17 @@ import BreakpointMgr from './BreakpointMgr';
 import DebuggerView from './DebuggerView';
 import path from 'path';
 
+
+if (!Promise.prototype.always) {
+    Promise.prototype.always = function(onResolveOrReject) {
+        return this.then(onResolveOrReject,
+            function(reason) {
+                onResolveOrReject(reason);
+                throw reason;
+            });
+        };
+}
+
 const breakpointMgr = new BreakpointMgr();
 
 function getRelativePath(fileName) {
@@ -16,8 +27,9 @@ function getRelativePath(fileName) {
         return '';
     }
 
-    if (options.cartridges && options.cartridges.length) {
+    if (options.cartridges && options.cartridges.trim().length) {
         let cartPath = options.cartridges
+            .split('\n')
             .map(str => str.trim())
             .find((path) => fileName.indexOf(path) !== -1);
 
@@ -38,8 +50,9 @@ function mapDwPathToProject(dwPath) {
     const workspacePath = atom.project.getPaths()[0];
     let newPath = dwPath.split('/').join(path.sep);
 
-    if (options.cartridges && options.cartridges.length) {
+    if (options.cartridges && options.cartridges.trim().length) {
         let cartPath = options.cartridges
+            .split('\n')
             .map(str => str.trim())
             .find((cartridge) => dwPath.indexOf(cartridge.split(path.sep).pop()) !== -1);
 
@@ -65,7 +78,6 @@ export default class Debugger {
             visible: false
         });
         this.debuggerView.on('close', this.deactivate.bind(this));
-        this.debuggerView.on('update', this.update.bind(this));
         this.debuggerView.on('stepover', this.stepover.bind(this));
         this.debuggerView.on('stop', this.stop.bind(this));
         this.debuggerView.on('resume', this.resume.bind(this));
@@ -146,23 +158,40 @@ export default class Debugger {
         }
     }
     update() {
-        this.connectionInstanse.getThreads().then((threads) => {
-            if (threads.length) {
-                threads.forEach((thread) => {
-                    // work only with one active thread, others just resume
-                    if (!this.currentThread || (this.currentThread && this.currentThread.threadInfo.id === thread.threadInfo.id)) {
-                        this.currentThread = thread;
-                    } else if (this.currentThread && this.currentThread.threadInfo.id !== thread.threadInfo.id) {
-                        thread.resume();
-                    }
-                });
-            } else {
+        if (this.isActive()) {
+            this.connectionInstanse.getThreads().then((threads) => {
+                if (threads.length) {
+                    threads.forEach((thread) => {
+                        // work only with one active thread, others just resume
+                        if (!this.currentThread || (this.currentThread && this.currentThread.threadInfo.id === thread.threadInfo.id)) {
+                            this.currentThread = thread;
+                        } else if (this.currentThread && this.currentThread.threadInfo.id !== thread.threadInfo.id) {
+                            thread.resume();
+                        }
+                    });
+                } else {
+                    this.currentThread = null;
+                }
+                if (
+                    !this.currentThread ||
+                    !window.currentThread ||
+                    window.currentThread.threadInfo.id !== this.currentThread.threadInfo.id ||
+                    window.currentThread.threadInfo.call_stack[0].location.line_number !== this.currentThread.threadInfo.call_stack[0].location.line_number
+                ) {
+                    this.debuggerView.updateThread(this.currentThread);
+                    this.setActiveBreakpoint(this.currentThread);
+                    window.currentThread = this.currentThread;
+                }
+
+            }).catch(err => {
                 this.currentThread = null;
-            }
-            this.debuggerView.updateThread(this.currentThread);
-            this.setActiveBreakpoint(this.currentThread);
-            window.currentThread = this.currentThread;
-        });
+                window.currentThread;
+                this.debuggerView.updateThread(null);
+                this.setActiveBreakpoint(null);
+                atom.notifications.addError('Keep alive: ' + err);
+            });
+
+        }
     }
     isActive() {
         return !!this.connectionInstanse;
@@ -180,6 +209,7 @@ export default class Debugger {
         }
         return this.connectionInstanse.init().then(() => {
             this.debuggerPanel.show();
+            this.timer = setInterval(this.update.bind(this), 10000);
         }).then(() => {
             atom.notifications.addInfo('Activated Debugger');
         }, (err) => {
@@ -195,6 +225,7 @@ export default class Debugger {
             return this.connectionInstanse.destroy().then(() => {
                 this.connectionInstanse = null;
                 this.debuggerPanel.hide();
+                clearInterval(this.timer);
             }).then(() => {
                 atom.notifications.addInfo('Dectivated Debugger');
             }, (err) => {
