@@ -1,11 +1,9 @@
 'use babel';
 
 /** @jsx elementCreator */
-import {domElement as elementCreator} from './elementCreator';
 import EventEmitter from 'events';
 
-const path = require('path'),
-      fs = require('fs');
+const path = require('path');
 
 export default class OptionsView {
     constructor (currentOptions = {}) {
