@@ -3,32 +3,40 @@
 
 export default class ProgressView {
 
-	constructor() {
-		this.el = document.createElement('div');
+    constructor() {
+        this.el = document.createElement('div');
 
-		this.el.className = 'b-bart';
-		const spin = document.createElement('span');
+        this.el.className = 'b-bart';
+        const spin = document.createElement('span');
 
-		spin.className = 'loading loading-spinner-tiny inline-block';
+        spin.className = 'loading loading-spinner-tiny inline-block';
 
-		const message = document.createElement('span');
+        const message = document.createElement('span');
 
-		message.textContent = 'bart';
+        message.textContent = 'bart';
 
-		this.msg = message;
+        this.msg = message;
 
-		this.el.appendChild(spin);
-		this.el.appendChild(message);
-	}
-	setMessage (msg) {
-		this.el.classList.remove('text-error');
-		this.msg.textContent = msg;
-	}
-	setError (msg) {
-		this.el.classList.add('text-error');
-		this.msg.textContent = msg;
-	}
-	getElement () {
-		return this.el;
-	}
+        this.el.appendChild(spin);
+        this.el.appendChild(message);
+    }
+    add(msg) {
+        this.setMessage(msg);
+    }
+    clean() {
+        this.setMessage('');
+    }
+    remove() {
+        // ????
+    }
+    setMessage (msg) {
+        this.el.classList.remove('text-error');
+        this.msg.textContent = msg;
+    }
+    setError () {
+        throw Error('Depricated!');
+    }
+    getElement () {
+        return this.el;
+    }
 }
