@@ -1,20 +1,11 @@
 'use babel'
 
-if (!Promise.prototype.always) {
-    Promise.prototype.always = function(onResolveOrReject) {
-        return this.then(onResolveOrReject,
-            function(reason) {
-                onResolveOrReject(reason);
-                throw reason;
-            });
-        };
-}
-
-
 import ProgressView from './ProgressView';
 import OptionsView from './OptionsView';
 import NewCartridgeView from './NewCartridgeView';
 import Debugger from './Debugger';
+const Observable = require('rxjs/Observable').Observable;
+require('rxjs/add/observable/concat');
 
 const atomP = require('atom'),
     path = require('path'),
@@ -25,7 +16,8 @@ const atomP = require('atom'),
 var watcher,
     debuggerInstance,
     timer,
-    repo;
+    repo,
+    uploadCartridgesSubscription;
 
 const path2DwPath = (fileName) => {
     const options = module.exports.getOptions();
@@ -64,7 +56,7 @@ const Bart = {
         if (typeof opts.watchlist === 'string' && opts.watchlist) {
             const chokidar = require('chokidar');
             watcher = chokidar.watch(null, {
-                ignored: /[\/\\]\./,
+                ignored: /[\/\\]\./, // eslint-disable-line no-useless-escape
                 persistent: true,
                 ignoreInitial: true,
                 followSymlinks: false
@@ -97,9 +89,8 @@ const Bart = {
             return;
         }
 
-
         timer = setInterval(() => {
-            if(atom.config.get('Bart.autoBuildEnabled')) {
+            if (atom.config.get('Bart.autoBuildEnabled')) {
                 const localRepo = getRepo();
                 if (localRepo && currentBranch !== localRepo.getShortHead()) {
                     currentBranch = localRepo.getShortHead();
@@ -107,7 +98,6 @@ const Bart = {
                 }
             }
         }, 5000);
-
     },
     setOptions : (options) => {
         if (options.saveconfigtofile) {
@@ -154,7 +144,7 @@ const Bart = {
         var configFilePath = '';
 
         configFilePath = (getPaths[0] ? getPaths[0] : './');
-        return configFilePath + (~configFilePath.indexOf('/') ? '/' : '\\') + fileName;
+        return configFilePath + (configFilePath.indexOf('/') ? '/' : '\\') + fileName;
     },
     getConfigFromFile : () => {
         const configFilePath = Bart.getConfigFilePath();
@@ -163,8 +153,8 @@ const Bart = {
         if (fs.existsSync(configFilePath)) {
             try {
                 confObject = JSON.parse(String(fs.readFileSync(configFilePath)));
-            } catch (e) {
-                console.warn('Parse dw.json error: ' + e);
+            } catch (err) {
+                console.warn('Parse dw.json error: ' + err);
             }
         }
 
@@ -180,9 +170,15 @@ const Bart = {
     },
     setConfigToFile : (options) => {
         const configFilePath = Bart.getConfigFilePath();
-        var configJson = JSON.stringify(options) || '';
+        var configJson = JSON.stringify(options, null, '    ') || '';
 
         fs.writeFileSync(configFilePath, configJson);
+    },
+    consumeSignal(registry) {
+      const provider = registry.create()
+      subscriptions.add(provider);
+
+      Bart.progressMsg = provider;
     },
     activate : () => {
         subscriptions.add(atom.commands.add('atom-workspace', {
@@ -204,21 +200,18 @@ const Bart = {
                 Bart.saveFile(params.path);
             });
         });
-        this.progressMsg = new ProgressView();
-        this.progressPanel = atom.workspace.addBottomPanel({
-            item: this.progressMsg.getElement(),
+        Bart.progressMsg = new ProgressView();
+
+        Bart.optionsView = new OptionsView(Bart.getOptions());
+        Bart.optionsPanel = atom.workspace.addModalPanel({
+            item: Bart.optionsView.getElement(),
             visible: false
         });
 
-        this.optionsView = new OptionsView(Bart.getOptions());
-        this.optionsPanel = atom.workspace.addModalPanel({
-            item: this.optionsView.getElement(),
-            visible: false
+        Bart.optionsView.on('cancel', () => {
+            Bart.optionsPanel.hide();
         });
-        this.optionsView.on('cancel', () => {
-            this.optionsPanel.hide();
-        });
-        this.optionsView.on('save', (options) => {
+        Bart.optionsView.on('save', (options) => {
             const WebDav = require('./webdav');
 
             if (!options.password) {
@@ -231,29 +224,37 @@ const Bart = {
             const webdav = new WebDav(options);
 
             if (options.version) {
-                webdav.dirList().then(() => {
+
+              const subscription = webdav.dirListRx().subscribe(
+                // next
+                () => {
                     Bart.setOptions(options);
-                    this.optionsView.clean();
-                    this.optionsPanel.hide();
+                    Bart.optionsView.clean();
+                    Bart.optionsPanel.hide();
                     Bart.initWatcher();
-                    this.optionsView.setError('');
+                    Bart.optionsView.setError('');
                     if (options.uploadaftersave) {
                         Bart.uploadAll();
                     }
-                }, (err) => {
-                    this.optionsView.setError(err);
-                });
+                },
+                // error
+                err => Bart.optionsView.setError(err),
+                // complite
+                () => subscription.unsubscribe()
+              );
             } else {
-                this.optionsView.setError('Code version must not be empty');
+                Bart.optionsView.setError('Code version must not be empty');
             }
         });
-        this.newCartridgeView = new NewCartridgeView();
-        this.newCartridgePanel = atom.workspace.addModalPanel({
-            item : this.newCartridgeView.getElement(),
+
+        Bart.newCartridgeView = new NewCartridgeView();
+        Bart.newCartridgePanel = atom.workspace.addModalPanel({
+            item : Bart.newCartridgeView.getElement(),
             visible : false
         });
-        this.newCartridgeView.on( 'cancel', () => {
-            this.newCartridgePanel.hide();
+
+        Bart.newCartridgeView.on( 'cancel', () => {
+            Bart.newCartridgePanel.hide();
         });
 
         this.newCartridgeView.on( 'create', ( options ) => {
@@ -262,7 +263,6 @@ const Bart = {
 
         Bart.initWatcher();
         Bart.initRepo();
-        console.log('bart alive!');
     },
     deactivate : () => {
         subscriptions.dispose();
@@ -272,6 +272,11 @@ const Bart = {
         if (timer) {
             clearInterval(timer);
         }
+
+        if (uploadCartridgesSubscription) {
+            uploadCartridgesSubscription.unsubscribe();
+        }
+
         if (debuggerInstance && debuggerInstance.isActive()) {
             debuggerInstance.deactivate();
             debuggerInstance = undefined;
@@ -297,17 +302,23 @@ const Bart = {
         const webdav = new WebDav(Object.assign({
             root: currentRoot
         }, options));
-        this.progressMsg.setMessage(`Uploading file: "${path.basename(fileName)}" ...`);
-        this.progressPanel.show();
 
-        webdav.post(fileName).then(() => {
-            this.progressPanel.hide();
-            console.log('Bart upload of ' + fileName + ' completed successfuly');
-        }, (err) => {
-            this.progressMsg.setError(err);
-            setTimeout(() => this.progressPanel.hide(), 10000);
-            atom.notifications.addError(err);
-        });
+        const msg = `Uploading file: "${path.basename(fileName)}" ...`
+        Bart.progressMsg.add(msg);
+
+        const subscription = webdav.postRx(fileName).subscribe(
+            () => Bart.progressMsg.remove(msg),
+            err => {
+                Bart.progressMsg.remove(msg);
+                const errorMsg = `<b>Error: ${err}</b>`;
+                Bart.progressMsg.add(errorMsg);
+                setTimeout(() => {
+                    Bart.progressMsg.remove(errorMsg);
+                }, 10000);
+                atom.notifications.addError(err);
+            },
+            () => subscription.unsubscribe()
+        );
     },
     uploadFile : (event) => {
         const _get = require('lodash/get'),
@@ -327,43 +338,67 @@ const Bart = {
         const WebDav = require('./webdav');
         const webdav = new WebDav(options);
 
-        this.progressPanel.show();
+        let prefMsg = null;
 
+
+        this.progressPanel.show();
         const notify = (msg) => {
-            console.info(msg);
-            this.progressMsg.setMessage(msg);
+            if (prefMsg) {
+                Bart.progressMsg.remove(prefMsg);
+            }
+            Bart.progressMsg.add(msg);
+            prefMsg = msg;
         }
-        const paths = atom.project.getDirectories().map((dir) => {
+
+        if (uploadCartridgesSubscription) {
+            Bart.progressMsg.clear();
+            uploadCartridgesSubscription.unsubscribe();
+        }
+        uploadCartridgesSubscription = Observable.concat(...atom.project.getDirectories().map(dir => {
             if (options.cartridges && options.cartridges.length) {
-                return options.cartridges
+
+                const toUpload = options.cartridges
                     .map(str => str.trim())
-                    .reduce( (promise, cartridge) => {
+                    .filter(Boolean)
+                    .reduce((promise, cartridge) => {
                         return promise.then(() => {
                             const dirToUpload = path.join(dir.getPath(), cartridge);
                             if (fs.existsSync(dirToUpload)) {
-                                return webdav.uploadCartridges(dirToUpload, notify, true);
+                                return webdav.uploadCartridgesRx(dirToUpload, notify, {isCartridge: true});
                             }
                         });
                     }, Promise.resolve());
+                return Observable.concat(...toUpload);
             } else {
                 return webdav.uploadCartridges(dir.getPath(), notify);
             }
-        });
-        Promise.all(paths).then(() => {
-            console.log('success');
-            this.progressPanel.hide();
-        }, (err) => {
-            this.progressMsg.setError(err);
-            atom.notifications.addError(err.toString());
-            setTimeout(() => this.progressPanel.hide(), 10000);
-        });
+        }))
+        .subscribe(
+          // next
+          () => {},
+          // error
+          err => {
+              Bart.progressMsg.clear();
+              const errorMsg = `Error: ${err}`;
+              Bart.progressMsg.add(errorMsg);
+              setTimeout(() => {
+                  Bart.progressMsg.remove(errorMsg);
+              }, 1000);
+              atom.notifications.addError(err.toString());
+          },
+          // complete
+          () => {
+              Bart.progressMsg.clear();
+              uploadCartridgesSubscription = null;
+          }
+      );
     },
     openConfig: () => {
-        if (this.optionsPanel.isVisible()) {
-            this.optionsPanel.hide();
+        if (Bart.optionsPanel.isVisible()) {
+            Bart.optionsPanel.hide();
         } else {
-            this.optionsPanel.show();
-            this.optionsView.displayAvailable();
+            Bart.optionsPanel.show();
+            Bart.optionsView.displayAvailable();
         }
     },
     toggleDebugger: () => {
@@ -388,10 +423,10 @@ const Bart = {
         }
     },
     openCreateNewCartridgeView : () => {
-        if (this.newCartridgePanel.isVisible()) {
-            this.newCartridgePanel.hide();
+        if (Bart.newCartridgePanel.isVisible()) {
+            Bart.newCartridgePanel.hide();
         } else {
-            this.newCartridgePanel.show();
+            Bart.newCartridgePanel.show();
         }
     },
     createDWRECartridge : (options) => {
@@ -411,17 +446,17 @@ const Bart = {
 
                         if (result) {
 
-                            this.newCartridgePanel.hide();
-                            this.newCartridgeView.clean();
+                            Bart.newCartridgePanel.hide();
+                            Bart.newCartridgeView.clean();
                             atom.notifications.addSuccess('Successfully created Cartridge ' + cartridgeName);
 
                             //If enabled, add the newly added cartridgePath to current ATOM workspace
-                            if(options.addtoproject) {
+                            if (options.addtoproject) {
                                 atom.project.addPath(completeCartridgePath);
                             }
 
                             //If enabled, link(upload) the newly created cartridge to DWRE server
-                            if(options.uploadaftercreate) {
+                            if (options.uploadaftercreate) {
                                 const bartOptions = Bart.getOptions();
                                 if (!bartOptions.password) {
                                     Bart.openConfig();
@@ -430,73 +465,83 @@ const Bart = {
                                 const WebDav = require('./webdav');
                                 const webdav = new WebDav(bartOptions);
 
-                                this.progressPanel.show();
-
+                                let prefMsg = null;
                                 const notify = (msg) => {
-                                    console.info(msg);
-                                    this.progressMsg.setMessage(msg);
+                                    if (prefMsg) {
+                                        Bart.progressMsg.remove(prefMsg);
+                                    }
+                                    Bart.progressMsg.add(msg);
+                                    prefMsg = msg;
                                 }
+
                                 //Upload the newly created empty Cartridge(DIR)
-                                webdav.uploadCartridges(completeCartridgePath, notify, true, true).then(()=> {
-                                    this.progressPanel.hide();
-                                    atom.notifications.addSuccess('Successfully uploaded Cartridge ' + cartridgeName);
-                                });
+                                webdav.uploadCartridgesRx(completeCartridgePath, notify, {isCartridge: true, isDirectory: true})
+                                .subscribe(
+                                    () => {},
+                                    err => {
+                                        atom.notifications.addError(err);
+                                    },
+                                    () => {
+                                        Bart.progressMsg.clear();
+                                        atom.notifications.addSuccess('Successfully uploaded Cartridge ' + cartridgeName);
+                                    }
+                                );
                             }
-                        }
-                        else {
+                        } else {
                             atom.notifications.addError('Error creating cartridge. Please see console log for more detailed message');
                         }
-                } catch(e) {
-                    console.log(e.stack);
-                    if(e.code === 'EEXIST') {
-                        atom.notifications.addError('Cartridge at given path already exists');
-                    }
-                    else if(e.code === 'ENOENT') {
-                        atom.notifications.addError('Directory Path does not exists');
-                    }
-                    else {
-                        atom.notifications.addError(e.message);
-                    }
+                } catch (err) {
+                      if (err.code === 'EEXIST') {
+                          atom.notifications.addError('Cartridge at given path already exists');
+                      }
+                      else if (err.code === 'ENOENT') {
+                          atom.notifications.addError('Directory Path does not exists');
+                      }
+                      else {
+                          atom.notifications.addError(err.message);
+                      }
                 }
             }
             else {
-                atom.notifications.addError("Invalid cartridgePath");
+                atom.notifications.addError('Invalid cartridgePath');
             }
         }
         else {
-            atom.notifications.addError("Invalid cartridgeName");
+            atom.notifications.addError('Invalid cartridgeName');
         }
     },
     makeDirectory : (completeCartridgePath, cartridgeName) => {
 
         let success = true;
-        directories_to_create = [
-            path.join(completeCartridgePath, "cartridge"),
-            path.join(completeCartridgePath, "cartridge", "forms"),
-            path.join(completeCartridgePath, "cartridge", "forms", "default"),
-            path.join(completeCartridgePath, "cartridge", "pipelines"),
-            path.join(completeCartridgePath, "cartridge", "scripts"),
-            path.join(completeCartridgePath, "cartridge", "static"),
-            path.join(completeCartridgePath, "cartridge", "static", "default"),
-            path.join(completeCartridgePath, "cartridge", "templates"),
-            path.join(completeCartridgePath, "cartridge", "templates", "default"),
-            path.join(completeCartridgePath, "cartridge", "templates", "resources"),
-            path.join(completeCartridgePath, "cartridge", "webreferences"),
-            path.join(completeCartridgePath, "cartridge", "webreferences2"),
-        ]
-        for (let index in directories_to_create)
-        {
-            try{
-                fs.mkdirSync(directories_to_create[index]);
-            }catch(e){
+        const directoriesToCreate = [
+            path.join(completeCartridgePath, 'cartridge'),
+            path.join(completeCartridgePath, 'cartridge', 'controllers'),
+            path.join(completeCartridgePath, 'cartridge', 'forms'),
+            path.join(completeCartridgePath, 'cartridge', 'forms', 'default'),
+            path.join(completeCartridgePath, 'cartridge', 'pipelines'),
+            path.join(completeCartridgePath, 'cartridge', 'scripts'),
+            path.join(completeCartridgePath, 'cartridge', 'static'),
+            path.join(completeCartridgePath, 'cartridge', 'static', 'default'),
+            path.join(completeCartridgePath, 'cartridge', 'templates'),
+            path.join(completeCartridgePath, 'cartridge', 'templates', 'default'),
+            path.join(completeCartridgePath, 'cartridge', 'templates', 'resources'),
+            path.join(completeCartridgePath, 'cartridge', 'webreferences'),
+            path.join(completeCartridgePath, 'cartridge', 'webreferences2'),
+        ];
+
+        directoriesToCreate.forEach(directory => {
+            try {
+                fs.mkdirSync(directory);
+            } catch (err) {
                 success = false;
-                console.log(e.message);
-                break;
             }
-        }
+        });
+
         if (success) {
-            const promiseArray = [Bart.createProjectPropertiesFile(completeCartridgePath, cartridgeName),
-            Bart.createCartridgePropertiesFile(completeCartridgePath, cartridgeName)]
+            const promiseArray = [
+                Bart.createProjectPropertiesFile(completeCartridgePath, cartridgeName),
+                Bart.createCartridgePropertiesFile(completeCartridgePath, cartridgeName)
+            ];
 
             Promise.all(promiseArray).then( () => {
                 console.log('success');
@@ -537,8 +582,8 @@ const Bart = {
 
         return new Promise( (resolve, reject) => {
 
-            fs.writeFile(completeProjectFilePath, projectFileContent, function(error){
-                if(error) {
+            fs.writeFile(completeProjectFilePath, projectFileContent, function(error) {
+                if (error) {
                     reject(error);
                 }
                 resolve("successfully added properties file");
@@ -547,22 +592,29 @@ const Bart = {
     },
     createCartridgePropertiesFile : (completeCartridgePath, cartridgeName) => {
 
-        const cartridgePropertiesFilePath = path.join(completeCartridgePath, 'cartridge', util.format('%s.properties', cartridgeName));
-        let cartridgePropertiesFileContent = `demandware.cartridges.%s.id=%s
-demandware.cartridges.%s.multipleLanguageStorefront=true`
-        cartridgePropertiesFileContent = util.format(cartridgePropertiesFileContent, cartridgeName, cartridgeName, cartridgeName);
+        const cartridgePropertiesFilePath = path.join(
+          completeCartridgePath,
+          'cartridge',
+          util.format('%s.properties', cartridgeName));
 
-        return new Promise( (resolve, reject) => {
-            fs.writeFile(cartridgePropertiesFilePath, cartridgePropertiesFileContent, function(error){
-                if(error) {
+        let cartridgePropertiesFileContent = `demandware.cartridges.%s.id=%s
+demandware.cartridges.%s.multipleLanguageStorefront=true`;
+
+        cartridgePropertiesFileContent = util.format(cartridgePropertiesFileContent,
+          cartridgeName, cartridgeName, cartridgeName);
+
+        return new Promise((resolve, reject) => {
+            fs.writeFile(cartridgePropertiesFilePath,
+              cartridgePropertiesFileContent, function(error) {
+                if (error) {
                     reject(error);
                 }
-                resolve("successfully added properties file");
+                resolve('successfully added properties file');
             });
         });
     },
     toggleAutoBuild : () => {
-        if(atom.config.get('Bart.autoBuildEnabled')) {
+        if (atom.config.get('Bart.autoBuildEnabled')) {
             atom.config.set('Bart.autoBuildEnabled', false);
             atom.notifications.addWarning('Successfully disabled auto build functionality');
         }
